Move ElementDescriptor to Pimcore\Model\Element namespace
