+++ doc/Development_Documentation/23_Installation_and_Upgrade/09_Upgrade_Notes/README.md
@@ -1,5 +1,10 @@
 # Upgrade Notes
 
+## 6.6.2
+- class `ElementDescriptor` has been moved from 'Pimcore\Model\Version' to 'Pimcore\Model\Element'.
+The BC layer will be removed in 7. Use the following [migration script](https://gist.github.com/weisswurstkanone/a63f733fe58930778f41c695f862724a) to migrate your version files
+if necessary.   
+
 ## 6.6.0
 - Default config for monolog handler `main` in prod environment is now `stream` instead of `fingers_crossed`. If you still want to use `fingers_crossed` please update your project config accordingly. 
 - `app` migration set is now located in `app/Migrations` instead of `app/Resources/migrations` - Pimcore will automatically move existing migration scripts for you (update your VCS!)
