+++ models/Element/ElementDescriptor.php
@@ -1,0 +1,80 @@
+<?php
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Enterprise License (PEL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PEL
+ */
+
+namespace Pimcore\Model\Element;
+
+class ElementDescriptor
+{
+    /** @var string */
+    protected $type;
+
+    /** @var int */
+    protected $id;
+
+    /**
+     * ElementDescriptor constructor.
+     *
+     * @param string $type
+     * @param int $id
+     */
+    public function __construct($type, $id)
+    {
+        $this->type = $type;
+        $this->id = $id;
+    }
+
+    /**
+     * @return string
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * @param string $type
+     *
+     * @return $this
+     */
+    public function setType($type)
+    {
+        $this->type = $type;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     *
+     * @return $this
+     */
+    public function setId(int $id)
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+}
+
+
+class_alias(ElementDescriptor::class, 'Pimcore\Model\Version\ElementDescriptor');
