+++ models/Version/ElementDescriptor.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Pimcore
  *
@@ -9,69 +8,28 @@
  * Full copyright and license information is available in
  * LICENSE.md which is distributed with this source code.
  *
- *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- *  @license    http://www.pimcore.org/license     GPLv3 and PEL
+ * @category   Pimcore
+ * @package    Element
+ *
+ * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
 namespace Pimcore\Model\Version;
 
-class ElementDescriptor
-{
-    /** @var string */
-    protected $type;
+@trigger_error(
+    'Pimcore\Model\Version\ElementDescriptor is deprecated since version 6.6.2 and will be removed in 7.0.0. ' .
+    ' Use ' . \Pimcore\Model\Element\ElementDescriptor::class . ' instead.',
+    E_USER_DEPRECATED
+);
 
-    /** @var int */
-    protected $id;
+class_exists(\Pimcore\Model\Element\ElementDescriptor::class);
 
+if (false) {
     /**
-     * ElementDescriptor constructor.
-     *
-     * @param string $type
-     * @param int $id
+     * @deprecated use \Pimcore\Model\Element\ElementDescriptor instead.
      */
-    public function __construct($type, $id)
+    class ElementDescriptor extends \Pimcore\Model\Element\ElementDescriptor
     {
-        $this->type = $type;
-        $this->id = $id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getType()
-    {
-        return $this->type;
-    }
-
-    /**
-     * @param string $type
-     *
-     * @return $this
-     */
-    public function setType($type)
-    {
-        $this->type = $type;
-
-        return $this;
-    }
-
-    /**
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return $this
-     */
-    public function setId(int $id)
-    {
-        $this->id = $id;
-
-        return $this;
     }
 }
